Docker for production
This completes much of the second half of the tutorial at https://londonappdeveloper.com/deploying-django-with-docker-compose/

The section for the development server is contained in the commit prior
to this one. Here, you can run the production services locally or on a
production server.

`docker-compose -f docker-compose-production.yml up`
`docker-compose -f docker-compose-production.yml down`
`docker-compose -f docker-compose-production.yml down --volumes` (to
destroy the volumes also)

The services included are for Django, postgresql and nginx. uswsgi is
used as the Django application server. Two separate scripts are used to
configure and run the proxy (nginx) and Django. Environment variables
are used to configure everything. These are contained in the .env file
which is kept out of the repository.